requirements.txt
Flask==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1

Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9

redis==5.0.7

pytest==8.3.2
requests==2.32.3
